Update packages.yml
wmctrl is a command that can be used to interact with an X Window manager that is compatible with the EWMH/NetWM specification. wmctrl can query the window manager for information, and it can request that certain window management actions be taken.

wmctrl is controlled entirely by its command line arguments. The command line arguments are used to specify the action to be performed (with options that modify behavior) and any arguments that might be needed to perform the actions.

The following sections define the supported actions and options. Arguments to the actions and options are written in the form <ARGNAME> in the descriptions below. The detailed syntax for writing arguments are in a single section dedicated to that purpose.